Update configuration handling and UI settings
Altered data types, from long to double, in temperature and SCR configuration files. Enhanced configuration handler with getter methods and a method for reading the configuration from disk. Features relating to SCR configuration (SCR type, voltage values, and current values) are now dynamically set based on the configuration. Adjusted interface (settings-view) to reflect these changes.

+++ src/main/java/de/bytestore/pvheating/objects/config/SCRConfig.java
@@ -7,16 +7,16 @@ import lombok.Data;
 public class SCRConfig {
     private SCRType type = SCRType.VOLTAGE;
 
-    private long minVoltage = 0;
+    private double minVoltage = 0;
 
-    private long maxVoltage = 10;
+    private double maxVoltage = 10;
 
-    private long minCurrent = 0;
+    private double minCurrent = 0;
 
-    private long maxCurrent = 20;
+    private double maxCurrent = 20;
 
-    private long minPWM = 5000;
+    private double minPWM = 5000;
 
-    private long maxPWM = 10000;
+    private double maxPWM = 10000;
 
 }
